add ReadSigmaVlqExt;
