+++ sigma-ser/src/zig_zag_encode.rs
